backend/requirements.txt
Django==3.2
django-filter==21.1
djangorestframework
drf-extra-fields==3.7.0
django-colorfield==0.10.1
gunicorn==20.1.0
psycopg2-binary==2.8.6
djoser
Pillow==9.2.0
python-decouple==3.5
drf-base64==2.0
webcolors==1.13
django-cors-headers==3.14.0
python-dotenv==0.21.1
python-decouple==3.5
